Never gonna make you cry

+++ src/main/java/com/bv/pet/jeduler/utils/AllowedAmount.java
@@ -1,15 +1,15 @@
 package com.bv.pet.jeduler.utils;
 
 public enum AllowedAmount {
-    CATEGORY((short) 20, "Category amount limit"),
-    TASK((short) 10000, "Task amount limit"),
-    SUBTASK((short) 20, "Subtask amount limit"),
-    USER((short) 1000, "User amount limit");
+    CATEGORY(20, "Category amount limit"),
+    TASK(10000, "Task amount limit"),
+    SUBTASK(20, "Subtask amount limit"),
+    USER(1000, "User amount limit");
 
-    final short amount;
+    final int amount;
     final String message;
 
-    AllowedAmount(short amount, String message) {
+    AllowedAmount(int amount, String message) {
         this.amount = amount;
         this.message = message;
     }
